simpify code structure
