+++ camel/messages/conversion/sharegpt/hermes/__init__.py
